In screen, bld ok
git-svn-id: https://subversion.assembla.com/svn/oddfry@13 572409ff-6c78-402f-b79f-49548ff2e957

+++ Code/OddFry/src/com/oddfry/globals/Globals.java
@@ -32,9 +32,9 @@ public class Globals {
 	 * Get current game time
 	 * @return current time in milliseconds
 	 */
-	public long getTime() {
+	public int getTime() {
 		//TODO game can slow down or speed up
-		return System.currentTimeMillis();
+		return (int) System.currentTimeMillis();
 	}
